+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.0] - 2022-02-14
+- Added bypass management pages (Thank you to our Sponsors!)
+  - Create, delete, edit, and view bypasses
+
 ## [1.9.0] - 2021-08-29
 - Reimplement splitpanes on agent page with a new library
 - Fix broken refresh button on agent list page